+++ backend/expenses/Touir.ExpensesManager.Expenses/Infrastructure/Options/EmailOptions.cs
@@ -1,0 +1,11 @@
+namespace Touir.ExpensesManager.Expenses.Infrastructure.Options
+{
+    public class EmailOptions
+    {
+        public string? Host { get; set; }
+        public int? Port { get; set; }
+        public string? Email { get; set; }
+        public string? Password { get; set; }
+        public bool EnableSsl { get; set; } = true;
+    }
+}